# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "clinicPlatform API",
    "version": "1.0.0",
    "description": "Auto-generated OpenAPI from Next app routes"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "paths": {
...
    "/api/clinics/{id}": {
      "get": {
        "summary": "Get clinic by id",
        "responses": {
          "200": {
            "description": "Clinic object",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "404": {
            "description": "Clinic not found"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": ""
          }
        ]
      },
      "put": {
        "summary": "Update clinic",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Clinic"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": ""
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete clinic",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Clinic"
                }
              }
            }
          },
          "400": {
            "description": "Cannot delete with dependencies"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": ""
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "address": {
                    "type": "string"
                  },
                  "city": {
                    "type": "string"
                  },
                  "state": {
                    "type": "string"
                  },
                  "zipCode": {
                    "type": "string"
                  },
                  "country": {
                    "type": "string"
                  },
                  "phone": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "website": {
                    "type": "string"
                  },
                  "logo": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "openingHours": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/clinics": {
      "get": {
        "summary": "List clinics",
        "responses": {
          "200": {
            "description": "Array of clinics",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ]
      },
      "post": {
        "summary": "Create clinic",
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "address": {
                    "type": "string"
                  },
                  "phone": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  }
                },
                "required": [
                  "name",
                  "address",
                  "phone",
                  "email"
                ]
              }
            }
          }
        }
      }
    },
...
    "/api/roles/{id}": {
      "put": {
        "summary": "PUT /api/roles/{id}",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Role"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": ""
          }
        ]
      },
      "delete": {
        "summary": "DELETE /api/roles/{id}",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Role"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": ""
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "emailVerified": {
            "type": "string",
            "format": "date-time"
          },
          "name": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/UserRole"
          },
          "image": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "clinicId": {
            "type": "string"
          },
          "clinic": {
            "$ref": "#/components/schemas/Clinic"
          },
          "patient": {
            "$ref": "#/components/schemas/Patient"
          },
          "appointments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Appointment"
            }
          },
          "consultations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Consultation"
            }
          },
          "prescriptions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Prescription"
            }
          },
          "accounts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Account"
            }
          },
          "sessions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Session"
            }
          },
          "notifications": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Notification"
            }
          },
          "auditLogs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AuditLog"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "email",
          "password",
          "appointments",
          "consultations",
          "prescriptions",
          "accounts",
          "sessions",
          "notifications",
          "auditLogs",
          "updatedAt"
        ]
      },
      "Clinic": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "logo": {
            "type": "string"
          },
          "settings": {
            "type": "object"
          },
          "users": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/User"
            }
          },
          "patients": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Patient"
            }
          },
          "appointments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Appointment"
            }
          },
          "services": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Service"
            }
          },
          "invoices": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Invoice"
            }
          },
          "notifications": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Notification"
            }
          },
          "auditLogs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AuditLog"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "name",
          "users",
          "patients",
          "appointments",
          "services",
          "invoices",
          "notifications",
          "auditLogs",
          "updatedAt"
        ]
      },
      "Patient": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "dateOfBirth": {
            "type": "string",
            "format": "date-time"
          },
          "gender": {
            "type": "string"
          },
          "bloodType": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "emergencyContact": {
            "type": "string"
          },
          "emergencyPhone": {
            "type": "string"
          },
          "medicalHistory": {
            "type": "string"
          },
          "allergies": {
            "type": "string"
          },
          "currentMedications": {
            "type": "string"
          },
          "clinicId": {
            "type": "string"
          },
          "clinic": {
            "$ref": "#/components/schemas/Clinic"
          },
          "appointments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Appointment"
            }
          },
          "consultations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Consultation"
            }
          },
          "prescriptions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Prescription"
            }
          },
          "invoices": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Invoice"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "firstName",
          "lastName",
          "dateOfBirth",
          "phone",
          "clinicId",
          "clinic",
          "appointments",
          "consultations",
          "prescriptions",
          "invoices",
          "updatedAt"
        ]
      },
      "Service": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "duration": {
            "type": "integer",
            "format": "int32"
          },
          "price": {
            "type": "number",
            "format": "float"
          },
          "active": {
            "type": "boolean"
          },
          "clinicId": {
            "type": "string"
          },
          "clinic": {
            "$ref": "#/components/schemas/Clinic"
          },
          "appointments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Appointment"
            }
          },
          "invoiceItems": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InvoiceItem"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "name",
          "duration",
          "price",
          "appointments",
          "invoiceItems",
          "updatedAt"
        ]
      },
      "Role": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "color": {
            "type": "string"
          },
          "permissions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RolePermission"
            }
          },
          "staffCount": {
            "type": "integer",
            "format": "int32"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "name",
          "permissions",
          "updatedAt"
        ]
      },
...
      "Appointment": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "startTime": {
            "type": "string",
            "format": "date-time"
          },
          "endTime": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "$ref": "#/components/schemas/AppointmentStatus"
          },
          "notes": {
            "type": "string"
          },
          "patientId": {
            "type": "string"
          },
          "patient": {
            "$ref": "#/components/schemas/Patient"
          },
          "doctorId": {
            "type": "string"
          },
          "doctor": {
            "$ref": "#/components/schemas/User"
          },
          "serviceId": {
            "type": "string"
          },
          "service": {
            "$ref": "#/components/schemas/Service"
          },
          "clinicId": {
            "type": "string"
          },
          "clinic": {
            "$ref": "#/components/schemas/Clinic"
          },
          "consultation": {
            "$ref": "#/components/schemas/Consultation"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "startTime",
          "endTime",
          "patientId",
          "patient",
          "doctorId",
          "doctor",
          "clinicId",
          "clinic",
          "updatedAt"
        ]
      },
...
      "Invoice": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "invoiceNumber": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/InvoiceStatus"
          },
          "totalAmount": {
            "type": "number",
            "format": "float"
          },
          "paidAmount": {
            "type": "number",
            "format": "float"
          },
          "dueDate": {
            "type": "string",
            "format": "date-time"
          },
          "notes": {
            "type": "string"
          },
          "paymentMethod": {
            "type": "string"
          },
          "paymentDate": {
            "type": "string",
            "format": "date-time"
          },
          "stripePaymentId": {
            "type": "string"
          },
          "patientId": {
            "type": "string"
          },
          "patient": {
            "$ref": "#/components/schemas/Patient"
          },
          "clinicId": {
            "type": "string"
          },
          "clinic": {
            "$ref": "#/components/schemas/Clinic"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InvoiceItem"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "invoiceNumber",
          "totalAmount",
          "patientId",
          "patient",
          "clinicId",
          "clinic",
          "items",
          "updatedAt"
        ]
      },
...
      "Notification": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "icon": {
            "type": "string"
          },
          "read": {
            "type": "boolean"
          },
          "userId": {
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "clinicId": {
            "type": "string"
          },
          "clinic": {
            "$ref": "#/components/schemas/Clinic"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "type",
          "title",
          "message",
          "userId",
          "user",
          "updatedAt"
        ]
      },
...
      "AuditLog": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "userEmail": {
            "type": "string"
          },
          "userRole": {
            "type": "string"
          },
          "action": {
            "type": "string"
          },
          "resource": {
            "type": "string"
          },
          "resourceId": {
            "type": "string"
          },
          "ipAddress": {
            "type": "string"
          },
          "userAgent": {
            "type": "string"
          },
          "method": {
            "type": "string"
          },
          "endpoint": {
            "type": "string"
          },
          "changes": {
            "type": "object"
          },
          "metadata": {
            "type": "object"
          },
          "status": {
            "type": "string"
          },
          "errorMessage": {
            "type": "string"
          },
          "clinicId": {
            "type": "string"
          },
          "clinic": {
            "$ref": "#/components/schemas/Clinic"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "action",
          "resource"
        ]
      }
    }
  }
}